more work

+++ dash-abci/src/execution/constants.rs
@@ -1,7 +1,7 @@
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 
-pub const EPOCHS_PER_YEAR: u64 = 20;
+pub const EPOCHS_PER_YEAR: u16 = 20;
 
 pub const EPOCHS_PER_YEAR_DEC: Decimal = dec!(20);
